app/requirements.txt
torchvision
mlflow
torch
matplotlib
tqdm
argparse
numpy
Pillow
pandas
fastapi
jinja2
uvicorn
python-multipart